backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-multipart==0.0.6
python-dotenv==1.0.0

# Google Cloud AI
google-cloud-aiplatform==1.38.1
google-generativeai==0.3.1
google-cloud-storage==2.13.0
google-cloud-documentai==2.20.2
google-cloud-videointelligence==2.14.0
google-auth==2.40.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# Database
sqlalchemy==2.0.23
neo4j==5.14.1

# ML/AI
sentence-transformers==2.2.2
faiss-cpu==1.8.0
langchain==0.0.340
langgraph==0.0.20
rank-bm25==0.2.2

# Document Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
pytesseract==0.3.10
easyocr==1.7.1
Pillow==10.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
httpx==0.25.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.0